requirements.txt
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.2.0
requests==2.22.0
scikit-image
scipy
image-slicer==0.3.0
PyYAML>=5.1
flask